requirements.txt
python-telegram-bot>=20.0,<21.0; python_version >= "3.11"
python-dotenv
